dsf_core::backend

Trait DatasetBackend

pub trait DatasetBackend {
    // Required methods
    fn get_metadata(&self, id: &str) -> BackendResult<MetaData>;
    fn mark_status(
        &self,
        id: &str,
        status: DataSetBusyStatus,
    ) -> BackendResult<()>;
    fn save_metadata(&self, metadata: &MetaData) -> BackendResult<()>;
    fn check_is_referenced(&self, target_id: &str) -> BackendResult<Vec<String>>;
    fn list_all_metadata(&self) -> BackendResult<Vec<MetaData>>;
    fn delete_metadata(&self, id: &str) -> BackendResult<()>;
}

Required Methods

fn get_metadata(&self, id: &str) -> BackendResult<MetaData>

Retrieves the corresponding metadata by the dataset ID.

fn mark_status(&self, id: &str, status: DataSetBusyStatus) -> BackendResult<()>

Mark MetaData status to ensure disk data and backend metadata consistency

fn save_metadata(&self, metadata: &MetaData) -> BackendResult<()>

Saves or updates the dataset metadata.

fn check_is_referenced(&self, target_id: &str) -> BackendResult<Vec<String>>

Checks if any datasets depend on the specified target_id.

Returns a list of dataset IDs that reference the target.

fn list_all_metadata(&self) -> BackendResult<Vec<MetaData>>

Lists all available dataset metadata from the backend.

fn delete_metadata(&self, id: &str) -> BackendResult<()>

Deletes the metadata associated with the specified dataset ID. note: this mucntion only deletes the metadata and detach this dataset from backend regisitration, real data on disk will be safe